chore: `x/subaccounts` move helper functions to to `lib` package (#1711)
